Refactor FolioReaderPage into domain-specific extensions and fix scroll scrubber logic
- Modularized FolioReaderPage.swift into separate extensions for Layout, Scroll, Progress, TOC, Highlights, Gestures, and Anchors.
- Resolved Swift 6 Sendable warnings in FolioReaderSharingProvider by applying @MainActor and restating conformance.
- Fixed ScrollScrubber scaling logic to prevent jumping to bottom and resolved Empty CFI errors by clamping offsets and correcting UISlider frame initialization.
- Improved thread safety in MockReadPositionProvider and corrected TextNode visibility logic in Bridge.js.
- Cleaned up accessibility modifiers in extensions from 'open' to 'public' to comply with Swift requirements.
- Expanded unit test suite for ReadPosition, Highlights, and JS Bridge.
