+++ .gemini/settings.json
@@ -1,0 +1,10 @@
+{
+  "mcpServers": {
+    "xcode-tools": {
+      "command": "xcrun",
+      "args": [
+        "mcpbridge"
+      ]
+    }
+  }
+}
