+++ Sources/FolioReaderKit/Page/FolioReaderPageDelegate.swift
@@ -1,0 +1,33 @@
+//
+//  FolioReaderPageDelegate.swift
+//  FolioReaderKit
+//
+
+import UIKit
+
+/// Protocol which is used from `FolioReaderPage`s.
+@objc public protocol FolioReaderPageDelegate: AnyObject {
+
+    /**
+     Notify that the page will be loaded. Note: The webview content itself is already loaded at this moment. But some java script operations like the adding of class based on click listeners will happen right after this method. If you want to perform custom java script before this happens this method is the right choice. If you want to modify the html content (and not run java script) you have to use `htmlContentForPage()` from the `FolioReaderCenterDelegate`.
+
+     - parameter page: The loaded page
+     */
+    @objc optional func pageWillLoad(_ page: FolioReaderPage)
+
+    /**
+     Notifies that page did load. A page load doesn't mean that this page is displayed right away, use `pageDidAppear` to get informed about the appearance of a page.
+
+     - parameter page: The loaded page
+     */
+    @objc optional func pageDidLoad(_ page: FolioReaderPage)
+    
+    /**
+     Notifies that page receive tap gesture.
+     
+     - parameter recognizer: The tap recognizer
+     */
+    @objc optional func pageTap(_ recognizer: UITapGestureRecognizer)
+    
+    @objc optional func pageStyleChanged(_ page: FolioReaderPage, _ reader: FolioReader)
+}
